Moving binding to separate folder.
MozReview-Commit-ID: CZ1WfQGH35n
