+++ gfx/moz.build
@@ -24,7 +24,7 @@ DIRS += [
     'config',
 ]
 
-EXPORTS += ['webrender/webrender.h']
+EXPORTS += ['webrender_bindings/src/bindings.h']
 
 if CONFIG['MOZ_ENABLE_SKIA']:
     DIRS += ['skia']
