+++ gfx/webrender/Cargo.toml
@@ -34,11 +34,6 @@ freetype = "0.1.1"
 [target.'cfg(target_os = "macos")'.dependencies]
 core-graphics = "0.4.1"
 core-text = "2.0"
-core-foundation = "0.2.2"
-
-[target.'cfg(target_os = "windows")'.dependencies]
-kernel32-sys = "0.2"
-winapi = "0.2.8"
 
 [profile.release]
 panic = "abort"
