+++ gfx/webrender/src/lib.rs
@@ -70,7 +70,6 @@ mod spring;
 mod texture_cache;
 mod tiling;
 mod util;
-pub mod bindings;
 
 mod platform {
     #[cfg(target_os="macos")]
@@ -94,15 +93,9 @@ pub mod renderer;
 extern crate core_graphics;
 #[cfg(target_os="macos")]
 extern crate core_text;
-#[cfg(target_os="macos")]
-extern crate core_foundation;
+
 #[cfg(not(target_os="macos"))]
 extern crate freetype;
-
-#[cfg(target_os="windows")]
-extern crate kernel32;
-#[cfg(target_os="windows")]
-extern crate winapi;
 
 extern crate app_units;
 extern crate bincode;
